// apps/admin/src/features/low-code/enums.ts
export enum LowCodeWidgetType {
  TEXT = 'TEXT',
  AUTO_COMPLETE = 'AUTO_COMPLETE',
  CHECKBOX = 'CHECKBOX',
  CASCADER = 'CASCADER',
  DATE_PICKER = 'DATE_PICKER',
  DATE_RANGE_PICKER = 'DATE_RANGE_PICKER',
  INPUT = 'INPUT',
  INPUT_NUMBER = 'INPUT_NUMBER',
  INPUT_TEXT_AREA = 'INPUT_TEXT_AREA',
  RADIO = 'RADIO',
  SELECT = 'SELECT',
  SLIDER = 'SLIDER',
  SWITCH = 'SWITCH',
  TIME_PICKER = 'TIME_PICKER',
  TREE_SELECT = 'TREE_SELECT',
  UPLOAD = 'UPLOAD',
  // 自定义控件
  LIST_SEARCH = 'LIST_SEARCH',
  TABLE_API = 'TABLE_API', // 表格接口
  CODE_BUTTON = 'CODE_BUTTON',
  API_BUTTON = 'API_BUTTON'
}